[Fix] circular dependency by moving `errorToMessage` to `Strings` namespace
