+++ src/base/common/error.ts
@@ -1,6 +1,5 @@
 import { IDisposable, toDisposable } from "src/base/common/dispose";
 import { Result, err, ok, panic } from "src/base/common/result";
-import { Arrays } from "src/base/common/utilities/array";
 import { mixin } from "src/base/common/utilities/object";
 import { Strings } from "src/base/common/utilities/string";
 
@@ -30,7 +29,7 @@ const _ErrorRegistrant = new class extends class ErrorRegistrant {
     constructor() {
         this._unexpectedErrorExternalCallback = (error: any) => {
             console.error(error);
-            panic(`on unexpected external error: ${errorToMessage(error)}`);
+            panic(`on unexpected external error: ${Strings.errorToMessage(error)}`);
         };
     }
 
@@ -144,55 +143,6 @@ export function isCancellationError(error: any): error is CancellationError {
 
 export function isExpectedError(error: any): error is ExpectedError {
     return error.type === ErrorType.Expected;
-}
-
-const UNKNOWN_MESSAGE = 'An unknown error occured. Please consult the log for more details.';
-
-/**
- * @description Try to convert an error to a human readable message in string.
- * @param error The given error.
- * @param verbose If output the stack trace.
- * @returns A string formated error message.
- * 
- * @note This function never throws.
- */
-export function errorToMessage(error: any, verbose: boolean = true): string {
-    if (!error) {
-        return UNKNOWN_MESSAGE;
-    }
-
-    if (Array.isArray(error)) {
-        const errors = Arrays.coalesce(error);
-        const firstErrorMessage = errorToMessage(errors[0], verbose);
-
-        if (errors.length > 1) {
-            return Strings.format('{0}, ({1} more errors in total)', [firstErrorMessage, errors.length - 1]);
-        }
-
-        return firstErrorMessage;
-    }
-
-    if (typeof error === 'string') {
-        return error;
-    }
-
-    if (error.stack && verbose) {
-        return Strings.format('{0} (stack trace - {1})', [error.message || UNKNOWN_MESSAGE, __stackToMessage(error.stack)]);
-    }
-
-    if (error.message) {
-		return error.message;
-	}
-
-    return `${UNKNOWN_MESSAGE}: ${JSON.stringify(error)}`;
-}
-
-function __stackToMessage(stack: any): string {
-    if (Array.isArray(stack)) {
-        return stack.join('\n');
-    } else {
-        return stack;
-    }
 }
 
 /**
