+++ src/platform/configuration/common/configurationModules/userConfiguration.ts
@@ -1,5 +1,5 @@
 import { Disposable, IDisposable } from "src/base/common/dispose";
-import { InitProtector, errorToMessage, tryOrDefault } from "src/base/common/error";
+import { InitProtector, tryOrDefault } from "src/base/common/error";
 import { AsyncResult, err, ok } from "src/base/common/result";
 import { Emitter, Event } from "src/base/common/event";
 import { DataBuffer } from "src/base/common/files/buffer";
