+++ src/workbench/services/shortcut/shortcutService.ts
@@ -16,9 +16,9 @@ import { ICommandService } from "src/platform/command/common/commandService";
 import { ContextKeyDeserializer } from "src/platform/context/common/contextKeyExpr";
 import { IRegistrantService } from "src/platform/registrant/common/registrantService";
 import { jsonSafeParse, jsonSafeStringify } from "src/base/common/json";
-import { errorToMessage } from "src/base/common/error";
 import { AsyncResult, Result, err, ok } from "src/base/common/result";
 import { FileOperationError } from "src/base/common/files/file";
+import { Strings } from "src/base/common/utilities/string";
 
 export const SHORTCUT_CONFIG_NAME = 'shortcut.config.json';
 export const IShortcutService = createService<IShortcutService>('shortcut-service');
@@ -241,7 +241,7 @@ export class ShortcutService extends Disposable implements IShortcutService {
                 commandArgs: undefined, // review
             });
         } catch (error) {
-            return err(new Error(errorToMessage(error)));
+            return err(new Error(Strings.errorToMessage(error)));
         }
 
         return ok();
